+++ src/pages/Course/Landing/ProgramBlocs/DecisionForm/Form/Form.module.scss
@@ -1,0 +1,69 @@
+
+@import 'styles/index';
+
+.form {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin-bottom: 0;
+}
+
+.inputWrap {
+  padding-bottom: set-calc(8);
+  display: flex;
+}
+
+.submitButton {
+  position: relative;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: set-calc(16) set-calc(16);
+  min-width: set-calc(58);
+
+  color: $white;
+
+  background-color: $green;
+  border-radius: 0 set-calc(15) set-calc(15) 0;
+  cursor: pointer;
+  pointer-events: all;
+
+  transition: background-color $d-hover ease-in-out, color ease-in-out $d-hover;
+
+  &:hover {
+    background-color: $green-hover;
+  }
+}
+
+.isSuccess {
+  pointer-events: none;
+}
+
+.isLoading {
+  opacity: 0.7;
+  pointer-events: none;
+}
+
+.isReset {
+  background-color: $yellow;
+  transition: background-color $d-hover ease-in-out, color ease-in-out $d-hover;
+
+  &:hover {
+    background-color: $yellow-hover;
+  }
+}
+
+.isDisabled {
+  background-color: $grey;
+  pointer-events: none;
+}
+
+.Error {
+  color: $orange;
+}
+
+.Success {
+  color: $green;
+}
